+++ package.json
@@ -16,15 +16,26 @@
   },
   "type": "module",
   "types": "dist/index.d.ts",
-  "main": "./dist/index.umd.cjs",
+  "main": "./dist/index.cjs",
   "module": "./dist/index.js",
   "source": "src/index.ts",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
       "import": "./dist/index.js",
-      "require": "./dist/index.umd.cjs"
-    }
+      "require": "./dist/index.cjs",
+      "types": "./dist/index.d.ts"
+    },
+    "./hooks": {
+      "import": "./dist/hooks/index.js",
+      "require": "./dist/hooks/index.cjs",
+      "types": "./dist/hooks/index.d.ts"
+    },
+    "./symbols": {
+      "import": "./dist/symbols/index.js",
+      "require": "./dist/symbols/index.cjs",
+      "types": "./dist/symbols/index.d.ts"
+    },
+    "./package.json": "./package.json"
   },
   "browser": "dist/index.js",
   "typings": "dist/index.d.ts",
