+++ src/hooks/index.ts
@@ -1,2 +1,10 @@
 export { useDrag } from './useDrag';
 export { useHoverIntent } from './useHoverIntent';
+export { useCameraControls } from './useCameraControls';
+export { useCenterGraph } from './useCenterGraph';
+export type {
+  CenterGraphInput,
+  CenterGraphOutput,
+  CenterNodesParams,
+  FitNodesParams
+} from './useCenterGraph';
