+++ src/hooks/useCameraControls.ts
