+++ src/hooks/useCenterGraph.ts
@@ -5,8 +5,11 @@ import { Vector3, Box3, PerspectiveCamera } from 'three';
 import { getLayoutCenter } from '../utils/layout';
 import { InternalGraphNode } from '../types';
 import { useStore } from '../store';
-import { isNodeInView, getDegreesToClosest2dAxis } from './utils';
-import { LayoutTypes } from 'layout/types';
+import {
+  isNodeInView,
+  getDegreesToClosest2dAxis
+} from '../CameraControls/utils';
+import { LayoutTypes } from '../layout/types';
 
 const PADDING = 50;
